+++ modules/devspec/desk/hyprland.nix
@@ -1,19 +1,55 @@
 # Hyprland
-{ lib, pkgs, ... }: {
-  hasGui = lib.mkDefault true;
-  hasDesktop = lib.mkDefault true;
+{
+  config,
+  lib,
+  pkgs,
+  ...
+}:
+let
+  cfg = config.programs.hyprland;
+  inherit (lib)
+    mkOption
+    types
+    ;
+in
+{
+  options = {
+    programs.hyprland.plugins.hyprexpo.enable = mkOption {
+      type = types.bool;
+      default = true;
+      description = "Set this to false to do not install Hyprexpo plugin for Hyprland";
+    };
+  };
 
-  xdg.portal.config.hyprland = { default = [ "hyprland" "gtk" ]; };
+  config = {
+    hasGui = lib.mkDefault true;
+    hasDesktop = lib.mkDefault true;
 
-  programs.hyprland.enable = true;
-  programs.hyprlock.enable = true;
-  services.hypridle.enable = true;
+    xdg.portal.config.hyprland = {
+      default = [
+        "hyprland"
+        "gtk"
+      ];
+    };
 
-  environment.systemPackages = with pkgs; [
-    hyprpaper
-    hyprpicker
-    hyprcursor
+    programs.hyprland.enable = true;
+    programs.hyprlock.enable = true;
+    services.hypridle.enable = true;
 
-    hyprlandPlugins.hyprexpo
-  ];
+    environment.systemPackages =
+      with pkgs;
+      [
+        hyprpaper
+        hyprpicker
+        hyprcursor
+      ]
+      ++ (
+        if cfg.plugins.hyprexpo.enable then
+          [
+            hyprlandPlugins.hyprexpo
+          ]
+        else
+          [ ]
+      );
+  };
 }